ML/requirements.txt
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
boilerpy3==1.0.7
cattrs==23.2.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.0
farm-haystack==1.25.2
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.51.0
fsspec==2024.3.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.22.2
hyperframe==5.2.0
idna==2.10
inflect==7.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy-imports==0.3.1
MarkupSafe==2.1.5
matplotlib==3.8.4
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.1
pillow==10.3.0
platformdirs==4.2.0
posthog==3.5.0
prompthub-py==4.0.0
pydantic==1.10.14
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
rfc3986==1.5.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.37.2
triton==2.2.0
typeguard==4.2.1
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
Werkzeug==3.0.2
